Revert old changes
