+++ examples/utils/example-utils/api-report/example-utils.api.md
@@ -176,7 +176,7 @@ export class MigrationTool extends DataObject implements IMigrationTool {
     // (undocumented)
     protected initializingFirstTime(): Promise<void>;
     // (undocumented)
-    get migrationState(): "migrated" | "migrating" | "collaborating" | "stopping";
+    get migrationState(): "collaborating" | "stopping" | "migrating" | "migrated";
     // (undocumented)
     get newContainerId(): string | undefined;
     // (undocumented)
@@ -250,7 +250,7 @@ export class SameContainerMigrationTool extends DataObject implements ISameConta
     // (undocumented)
     protected initializingFirstTime(): Promise<void>;
     // (undocumented)
-    get migrationState(): "migrated" | "collaborating" | "proposingMigration" | "stoppingCollaboration" | "proposingV2Code" | "waitingForV2ProposalCompletion" | "readyForMigration";
+    get migrationState(): "collaborating" | "migrated" | "proposingMigration" | "stoppingCollaboration" | "proposingV2Code" | "waitingForV2ProposalCompletion" | "readyForMigration";
     // (undocumented)
     get proposedVersion(): string | undefined;
     // (undocumented)
