+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -179,7 +179,6 @@ export {
 	SchemaBuilderOptions,
 	normalizeAllowedTypes,
 	normalizeField,
-	structuralName,
 } from "./schemaBuilderBase";
 export { SchemaBuilderInternal } from "./schemaBuilder";
